add loadbalancer_utils.go file
